[INJIVER-139] Added a separate reducer for verification Signed-off-by: Shiva <shivakumar@beehyv.com>
