+++ inji-verify/src/components/Home/VerificationSection/QrScanner.tsx
@@ -3,7 +3,7 @@ import {Scanner} from '@yudiel/react-qr-scanner';
 import CameraAccessDenied from "./CameraAccessDenied";
 import {ScanSessionExpiryTime} from "../../../utils/config";
 import {useAppDispatch} from "../../../redux/hooks";
-import {goHomeScreen, raiseAlert, verificationInit} from "../../../redux/features/verification/verificationSlice";
+import {goHomeScreen, raiseAlert, verificationInit} from "../../../redux/features/verification/verification.slice";
 
 let timer: NodeJS.Timeout;
