+++ inji-verify/src/components/Home/VerificationSection/UploadQrCode.tsx
@@ -2,7 +2,7 @@ import {scanFilesForQr} from "../../../utils/qr-utils";
 import {AlertMessages, UploadFileSizeLimits} from "../../../utils/config";
 import {ReactComponent as UploadIcon} from "../../../assets/upload-icon.svg";
 import {useAppDispatch} from "../../../redux/hooks";
-import {goHomeScreen, qrReadInit, raiseAlert, verificationInit} from "../../../redux/features/verification/verificationSlice";
+import {goHomeScreen, qrReadInit, raiseAlert, verificationInit} from "../../../redux/features/verification/verification.slice";
 
 function UploadButton({ displayMessage }: {displayMessage: string}) {
     return (
