+++ inji-verify/src/redux/store.ts
@@ -1,12 +1,16 @@
-import {configureStore} from "@reduxjs/toolkit";
-import verificationFlowReducer, {PreloadedState} from './features/verification/verificationSlice';
-import verificationSaga from './features/verification/verificationSaga';
+import {combineReducers, configureStore} from "@reduxjs/toolkit";
+import verificationFlowReducer, {PreloadedState} from './features/verification/verification.slice';
+import verificationSaga from './features/verification/verification.saga';
 import createSagaMiddleware from "redux-saga";
 
 const sagaMiddleware = createSagaMiddleware();
 
+const rootReducer = combineReducers({
+    verification: verificationFlowReducer
+})
+
 const store = configureStore({
-        reducer: verificationFlowReducer,
+        reducer: rootReducer,
         middleware: (getDefaultMiddleware) => getDefaultMiddleware({ thunk: false }).concat(sagaMiddleware),
         devTools: true/*process.env.NODE_ENV !== 'production'*/,
         preloadedState: PreloadedState as any
